build: update pinned versions of lightning-onion and btcwallet
In this commit, we update the pinned versions of lightning-onion as
well as btcwallet.

The newest lightning-onion instance adds some logging statements when
garbage collecting expired shared secret hashes.

The new btcwallet version fixes a recently introduced bug which would
cause us to remove all unman transactions on restart if the full node
new of any of them. Instead, we’ll now only remove these unmanned
transactions if it’s reported that the transactions were double spends,
or orphans.